some reviews

+++ repositories/repository.go
@@ -9,7 +9,7 @@ type BetRepository interface {
 	BetById(bet_id string) (models.Bet, error)
 	Bets(startIndex, perpage int64) ([]models.Bet, error)
 	RunningBets(startIndex, perpage int64) ([]models.Bet, error)
-	TotalRunningBetsMoney() float32
+	TotalRunningBetsMoney() float64
 	ProcessWin(amount float64, user_id string)
 }
